Correct RunenSpatial address and frame contracts (#9)
Implement the accepted RS3 foundational spatial contracts from issue #8.

- migrate stable spatial coordinates to signed i64
- add checked spatial errors, configuration, frames, hierarchy, clipmap, and ring arithmetic
- propagate checked failures atomically through demand and streaming
- preserve accepted demand and lifecycle behavior with restored regression coverage
- validate adapter, chunk-origin identity, hierarchy boundaries, Serde invariants, and documentation

Reviewed head: d618f0002958a736ed365fbf848b5eaf6884235e
Accepted base: 363d3ef515fbd60f3d758f06d2f88d7f06ef137f

+++ adapters/godot_world_streaming/src/bridge.rs
@@ -2,24 +2,24 @@ use godot::builtin::Vector3;
 use runen_spatial::{ChunkCoord3, ChunkId, WorldId};
 use runen_spatial_streaming::{ProviderEvent, ProviderEventKind, StreamRequestId};
 
-pub fn vector3_to_meters(value: Vector3) -> [f32; 3] {
-    [value.x, value.y, value.z]
+pub fn vector3_to_meters(value: Vector3) -> [f64; 3] {
+    [f64::from(value.x), f64::from(value.y), f64::from(value.z)]
 }
 
-pub fn chunk_coord_from_xyz(x: i32, y: i32, z: i32) -> ChunkCoord3 {
+pub fn chunk_coord_from_xyz(x: i64, y: i64, z: i64) -> ChunkCoord3 {
     ChunkCoord3 { x, y, z }
 }
 
-pub fn chunk_id_from_xyz(world_id: u16, x: i32, y: i32, z: i32) -> ChunkId {
+pub fn chunk_id_from_xyz(world_id: u16, x: i64, y: i64, z: i64) -> ChunkId {
     ChunkId::new(WorldId(world_id), chunk_coord_from_xyz(x, y, z))
 }
 
 pub fn provider_event_from_godot(
     world_id: u16,
     request_id: i64,
-    x: i32,
-    y: i32,
-    z: i32,
+    x: i64,
+    y: i64,
+    z: i64,
     kind: ProviderEventKind,
 ) -> Option<ProviderEvent> {
     let request_id = u64::try_from(request_id).ok()?;
